Extracting constants in test cases, some files missed (hymeko_daemon)

+++ hymeko_core/tests/minimal_tests/test_import/test_import_graphs.rs
@@ -31,22 +31,22 @@ mod test_import_graphs
         let sid_operator = interner.intern("operator");
 
         let did_elements = *idx.by_path.get(&PathKey(vec![sid_elements])).expect("elements missing");
-        assert_eq!(ir.decl_nodes[did_elements.0 as usize].kind, DeclKind::Node);
+        assert_eq!(ir.decl_nodes[did_elements.0].kind, DeclKind::Node);
 
         let did_operand = *idx.by_path.get(&PathKey(vec![sid_elements, sid_operand])).expect("operand missing");
         let did_operand2 = *idx.by_path.get(&PathKey(vec![sid_elements, sid_operand2])).expect("operand2 missing");
-        assert_eq!(ir.decl_nodes[did_operand.0 as usize].kind, DeclKind::Node);
-        assert_eq!(ir.decl_nodes[did_operand2.0 as usize].kind, DeclKind::Node);
+        assert_eq!(ir.decl_nodes[did_operand.0].kind, DeclKind::Node);
+        assert_eq!(ir.decl_nodes[did_operand2.0].kind, DeclKind::Node);
 
         let did_operator = *idx.by_path.get(&PathKey(vec![sid_elements, sid_operator])).expect("operator edge missing");
-        assert_eq!(ir.decl_nodes[did_operator.0 as usize].kind, DeclKind::Edge);
+        assert_eq!(ir.decl_nodes[did_operator.0].kind, DeclKind::Edge);
 
-        let edge_id = ir.decl_to_edge[did_operator.0 as usize].expect("operator not lowered as edge");
-        let edge_rec = &ir.edges[edge_id.0 as usize];
+        let edge_id = ir.decl_to_edge[did_operator.0].expect("operator not lowered as edge");
+        let edge_rec = &ir.edges[edge_id.0];
         assert_eq!(edge_rec.arcs.len(), 1, "operator edge should contain one arc");
 
         let arc_id = edge_rec.arcs[0];
-        let arc = &ir.arcs[arc_id.0 as usize];
+        let arc = &ir.arcs[arc_id.0];
         assert_eq!(arc.refs.len(), 2, "operator arc should connect two refs");
 
         match (&arc.refs[0], &arc.refs[1]) {
@@ -89,7 +89,7 @@ mod test_import_graphs
             .get(&PathKey(vec![ns, elements, operand]))
             .expect("missing basic_library.elements.operand in global index");
 
-        let kind = compiled.ir.decl_nodes[did.0 as usize].kind;
+        let kind = compiled.ir.decl_nodes[did.0].kind;
         assert!(matches!(kind, DeclKind::Node|DeclKind::Edge|DeclKind::HyperArc));
 
         let mut referenced = false;
@@ -107,7 +107,7 @@ mod test_import_graphs
         }
         assert!(referenced, "expected at least one arc ref to target basic_library.elements.operand");
         assert!(
-            compiled.ir.decl_hash.get(did.0 as usize).and_then(|x| *x).is_some(),
+            compiled.ir.decl_hash.get(did.0).and_then(|x| *x).is_some(),
             "expected decl hash for operand to be computed"
         );
